pyproject.toml
[project]
name = "brunel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "platformdirs>=4.9.6",
    "prompt-toolkit>=3.0.52",
    "pydantic-ai-slim[openai]>=1.78.0",
    "pydantic-settings>=2.14.0",
    "pytest>=9.0.3",
    "typer>=0.24.1",
]

[project.scripts]
brunel = "brunel.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brunel"]
